requirements.txt
viam-sdk==0.41.0
typing-extensions
pillow